claude_exp/OneDriveUpload/requirements.txt
Office365-REST-Python-Client>=2.5.0
msal>=1.28.0
pytest>=8.0.0
